requirements.txt
flask
slackclient
slackeventsapi